+++ nuxeo/groups.py
@@ -1,17 +1,11 @@
 # coding: utf-8
-from __future__ import unicode_literals
+from typing import TYPE_CHECKING, Dict, Optional
 
 from .endpoint import APIEndpoint
 from .models import Group
 
-try:
-    from typing import TYPE_CHECKING
-
-    if TYPE_CHECKING:
-        from typing import Dict, Optional, Text
-        from .client import NuxeoClient
-except ImportError:
-    pass
+if TYPE_CHECKING:
+    from .client import NuxeoClient
 
 
 class API(APIEndpoint):
@@ -20,19 +14,19 @@ class API(APIEndpoint):
     __slots__ = ("params",)
 
     def __init__(self, client, endpoint="group", headers=None):
-        # type: (NuxeoClient, Text, Optional[Dict[Text, Text]]) -> None
+        # type: (NuxeoClient, str, Optional[Dict[str, str]]) -> None
         self.params = {"fetch-group": ["memberUsers", "memberGroups"]}
-        super(API, self).__init__(client, endpoint=endpoint, cls=Group, headers=headers)
+        super().__init__(client, endpoint=endpoint, cls=Group, headers=headers)
 
     def get(self, group_id=None):
-        # type: (Optional[Text]) -> Group
+        # type: (Optional[str]) -> Group
         """
         Get the detail of a group.
 
         :param group_id: the id of the group
         :return: the group
         """
-        return super(API, self).get(path=group_id, params=self.params)
+        return super().get(path=group_id, params=self.params)
 
     def post(self, group):
         # type: (Group) -> Group
@@ -42,7 +36,7 @@ class API(APIEndpoint):
         :param group: the group to create
         :return: the created group
         """
-        return super(API, self).post(resource=group, params=self.params)
+        return super().post(resource=group, params=self.params)
 
     create = post  # Alias for clarity
 
@@ -54,13 +48,13 @@ class API(APIEndpoint):
         :param group: the group to update
         :return: the updated group
         """
-        return super(API, self).put(group)
+        return super().put(group)
 
     def delete(self, group_id):
-        # type: (Text) -> None
+        # type: (str) -> None
         """
         Delete a group.
 
         :param group_id: the id of the group to delete
         """
-        super(API, self).delete(group_id)
+        super().delete(group_id)