+++ nuxeo/users.py
@@ -1,17 +1,11 @@
 # coding: utf-8
-from __future__ import unicode_literals
+from typing import TYPE_CHECKING, Dict, Optional
 
 from .endpoint import APIEndpoint
 from .models import User
 
-try:
-    from typing import TYPE_CHECKING
-
-    if TYPE_CHECKING:
-        from typing import Dict, Optional, Text
-        from .client import NuxeoClient
-except ImportError:
-    pass
+if TYPE_CHECKING:
+    from .client import NuxeoClient
 
 
 class API(APIEndpoint):
@@ -20,18 +14,18 @@ class API(APIEndpoint):
     __slots__ = ()
 
     def __init__(self, client, endpoint="user", headers=None):
-        # type: (NuxeoClient, Text, Optional[Dict[Text, Text]]) -> None
-        super(API, self).__init__(client, endpoint=endpoint, cls=User, headers=headers)
+        # type: (NuxeoClient, str, Optional[Dict[str, str]]) -> None
+        super().__init__(client, endpoint=endpoint, cls=User, headers=headers)
 
     def get(self, user_id=None):
-        # type: (Optional[Text]) -> User
+        # type: (Optional[str]) -> User
         """
         Get the detail of a user.
 
         :param user_id: the id of the user
         :return: the user
         """
-        return super(API, self).get(path=user_id)
+        return super().get(path=user_id)
 
     def post(self, user):
         # type: (User) -> User
@@ -41,7 +35,7 @@ class API(APIEndpoint):
         :param user: the user to create
         :return: the created user
         """
-        return super(API, self).post(user)
+        return super().post(user)
 
     create = post  # Alias for clarity
 
@@ -53,16 +47,16 @@ class API(APIEndpoint):
         :param user: the user to update
         :return: the updated user
         """
-        return super(API, self).put(user)
+        return super().put(user)
 
     def delete(self, user_id):
-        # type: (Text) -> None
+        # type: (str) -> None
         """
         Delete a user.
 
         :param user_id: the id of the user to delete
         """
-        super(API, self).delete(user_id)
+        super().delete(user_id)
 
     def current_user(self):
         # type: () -> User